dont highlight match if value empty
This is a temporary fix that we need to prevent ui-select from
pre-selecting a value in a single select when the user has not
entered any text into the field and ngModel is empty.

This fix started out as a feature, but ended up being more like a
bug fix and needs to be refactored.  Also, this commit broke a few
tests, though it likely did so because the test is wrong now,
rather than the code not performing as expected.  Either way, the
tests need to be repaired and a pull request made to the origin
when the time comes.